design/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ectf25_design"
version = "2025.0+example"
requires-python = ">=3.11"

# TODO: add your custom dependencies to this list
dependencies = [
    "loguru",
    "pycryptodome",
    "argon2-cffi",
]

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''